% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_identifier.R
\name{check_identifier}
\alias{check_identifier}
\title{Check Identifier}
\usage{
check_identifier(df, x)
}
\arguments{
\item{dataframe}{The original data from the user.}

\item{identifier}{A unique identifier to split the data by, such as a customer ID, product ID, etc.}
}
\value{
Nothing if valid, an error message if an invalid input.
}
\description{
Helper function for fill_dates. Ensures that the identifier variable is a valid input from the user.
}
