% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dates.R
\name{fill_dates}
\alias{fill_dates}
\title{Fill Dates}
\usage{
fill_dates(
  dataframe,
  identifier = NA,
  date_col,
  min_date,
  max_date,
  time_sequence,
  fill_data = NA
)
}
\arguments{
\item{dataframe}{The original data with missing entries for certain dates.}

\item{identifier}{A unique identifier to split the data by, such as a customer ID, product ID, etc.}

\item{date_col}{The date column of the dataframe.}

\item{min_date}{The minimum date in the range that you want to fill.}

\item{max_date}{The maximum date in the range that you want to fill.}

\item{time_sequence}{The interval between dates, either "day", "week", "month", "quarter", or "year".}

\item{fill_data}{The data that should be filled for missing values. Default are NAs.}
}
\value{
A dataframe with missing entries filled.
}
\description{
Fill missing date values in a dataframe.
}
