% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_read_csv.R
\name{filter_read_csv}
\alias{filter_read_csv}
\title{Filter Read CSV}
\usage{
filter_read_csv(path, batch_size = 100, filters)
}
\arguments{
\item{path}{A file path to a csv}

\item{batch_size}{The size of each csv chunk to be read in}

\item{filters}{A vector of filters to be applied. Each filter should start with the column name and NOT the name of the file}
}
\value{
tibble
}
\description{
Returns a filtered tibble.
}
