% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{find_mode}
\alias{find_mode}
\title{Find Mode}
\usage{
find_mode(vector, remove_na = TRUE)
}
\arguments{
\item{vector}{A file path to a local csv file.}

\item{remove_na}{Whether to remove missing values from the sequence prior to finding the mode. Default is true.}
}
\value{
The most frequent value in the sequence.
}
\description{
Returns the mode of a given series.
}
