% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_lengths.R
\name{list_lengths}
\alias{list_lengths}
\title{List Lengths}
\usage{
list_lengths(x)
}
\arguments{
\item{x}{a vector of lists}
}
\value{
vector
}
\description{
Returns a vector of the lengths of each element in the vector.
}
