% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_read_csv.R
\name{filter_read_csv}
\alias{filter_read_csv}
\title{Filter Read CSV}
\usage{
filter_read_csv(path, batch_size, cols, filters)
}
\arguments{
\item{path}{A file path to a local csv file.}

\item{batch_size}{The size of each csv chunk to be read in. If not specified, then the data will be
broken into 10 chunks by default.}

\item{filters}{vector of filters to be applied. Each filter should start with the column name and
NOT the name of the file; filter_read_csv will take care of this automatically. For example, if one
of the filters was to be applied on the column "Age", the filter would be included in the vector
as "Age < 40", NOT as "df$AGE < 40".}
}
\value{
A filtered tibble.
}
\description{
Iteratively reads in and filters a CSV file based on specified filters.
}
